pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
]
